fix(dawcore): address all code review issues
- Add generation counter to prevent stale rAF callbacks on rapid
  disconnect/reconnect cycles
- Extract TrackDescriptor/ClipDescriptor to src/types.ts, bringing
  daw-editor.ts from 800 to 779 lines (headroom for future additions)
- Restore two-line @property style for consistency
- Fix spec doc: hostConnected fires BEFORE willUpdate (not after)
- Add test for stale rAF generation counter
- Re-export types from index.ts for backwards compatibility
