+++ packages/dawcore/src/elements/daw-editor.ts
@@ -1,6 +1,7 @@
 import { LitElement, html, css } from 'lit';
 import { customElement, property, state } from 'lit/decorators.js';
 import type { ClipTrack, FadeType, Peaks, PeakData } from '@waveform-playlist/core';
+import type { TrackDescriptor, ClipDescriptor } from '../types';
 import { createClipFromSeconds, createTrack, clipPixelWidth } from '@waveform-playlist/core';
 import { PeakPipeline } from '../workers/peakPipeline';
 import type { DawTrackElement } from './daw-track';
@@ -20,28 +21,6 @@ import type {
   LoadFilesResult,
 } from '../events';
 import { loadFiles as loadFilesImpl } from '../interactions/file-loader';
-
-export interface TrackDescriptor {
-  name: string;
-  src: string;
-  volume: number;
-  pan: number;
-  muted: boolean;
-  soloed: boolean;
-  clips: ClipDescriptor[];
-}
-
-interface ClipDescriptor {
-  src: string;
-  start: number;
-  duration: number;
-  offset: number;
-  gain: number;
-  name: string;
-  fadeIn: number;
-  fadeOut: number;
-  fadeType: FadeType;
-}
 
 @customElement('daw-editor')
 export class DawEditorElement extends LitElement {
@@ -84,7 +63,8 @@ export class DawEditorElement extends LitElement {
   private _trackElements = new Map<string, DawTrackElement>();
   private _childObserver: MutationObserver | null = null;
   private _audioResume = new AudioResumeController(this);
-  @property({ attribute: 'eager-resume' }) eagerResume?: string;
+  @property({ attribute: 'eager-resume' })
+  eagerResume?: string;
   private _pointer = new PointerHandler(this);
   private _viewport = (() => {
     const v = new ViewportController(this);
@@ -228,7 +208,6 @@ export class DawEditorElement extends LitElement {
     if (changedProperties.has('eagerResume')) {
       this._audioResume.target = this.eagerResume;
     }
-
     // Re-extract peaks at new zoom level from cached WaveformData (near-instant)
     if (changedProperties.has('samplesPerPixel') && this._clipBuffers.size > 0) {
       const reextracted = this._peakPipeline.reextractPeaks(
