+++ packages/dawcore/src/index.ts
@@ -27,6 +27,7 @@ export { DawTrackControlsElement } from './elements/daw-track-controls';
 
 export { AudioResumeController } from './controllers/audio-resume-controller';
 
+export type { TrackDescriptor, ClipDescriptor } from './types';
 export type { PointerEngineContract } from './interactions/pointer-handler';
 
 export type {
