+++ packages/dawcore/src/types.ts
@@ -1,0 +1,23 @@
+import type { FadeType } from '@waveform-playlist/core';
+
+export interface TrackDescriptor {
+  name: string;
+  src: string;
+  volume: number;
+  pan: number;
+  muted: boolean;
+  soloed: boolean;
+  clips: ClipDescriptor[];
+}
+
+export interface ClipDescriptor {
+  src: string;
+  start: number;
+  duration: number;
+  offset: number;
+  gain: number;
+  name: string;
+  fadeIn: number;
+  fadeOut: number;
+  fadeType: FadeType;
+}
